requirements-test.txt
# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
factory-boy==3.3.0
faker==20.1.0

# Testing utilities
coverage==7.3.2
responses==0.24.1
freezegun==1.2.2

# Database testing
pytest-postgresql==5.0.0
sqlalchemy-utils==0.41.1

